+++ thesisbackend/src/main/java/com/thesis/backend/model/entity/ml/MachineLearningOutput.java
@@ -16,15 +16,15 @@ public class MachineLearningOutput {
     @Column(name = "id")
     private long id;
     @Column(name = "x_start")
-    private int xStart;
+    private float xStart;
     @Column(name = "y_start")
-    private int yStart;
+    private float yStart;
     @Column(name = "floor_start")
     private int floorStart;
     @Column(name = "x_end")
-    private int xEnd;
+    private float xEnd;
     @Column(name = "y_end")
-    private int yEnd;
+    private float yEnd;
     @Column(name = "floor_end")
     private int floorEnd;
     @Column(name = "type")
